Add new adventure: Resistors in series
Add new adventure, "Resistors in series", to simulate and build a
circuit with resistors connected in series, compute their equivalent
resistance, and determine voltage across each resistor.

Also fix a few typos in "Ohm's law", and include an explicit section on
how to save and load a circuit in Circuit Simulator.

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>